// packages/backend/src/core/exchanges/okx/types/exchange/trade/place-limit-order/place-limit-order-request.interface.ts
import { OrderSide } from 'src/core/exchanges/okx/types/exchange/trade/common/types/order-side.type';

export interface IPlaceLimitOrderRequest {
  /**
   * Instrument ID, e.g `ADA-USDT`.
   */
  symbol: string;
  /**
   * Client-supplied order ID
   */
  clientOrderId: string;
  side: OrderSide;
  /**
   * Quantity to buy or sell.
   */
  quantity: number;
  /**
   * Order price.
   */
  price: number;
}